refactor(self_improve): split tests_cycle.rs (#1266)
tests_cycle.rs 814 -> 383 LOC; new tests_cycle_more.rs (~231) + tests_cycle_most.rs (~250).
