+++ src/self_improve/tests_cycle.rs
@@ -380,435 +380,3 @@ fn decide_with_zero_scores() {
         "zero net improvement should revert"
     );
 }
-
-#[test]
-fn decide_with_multiple_severe_regressions() {
-    let cfg = ImprovementConfig::default();
-    let baseline = make_score(0.70);
-    let post = make_score(0.80);
-    let regressions = vec![
-        Regression {
-            dimension: "specificity".into(),
-            baseline_score: 0.7,
-            current_score: 0.5,
-            delta: -0.20,
-            severity: RegressionSeverity::Severe,
-        },
-        Regression {
-            dimension: "temporal_awareness".into(),
-            baseline_score: 0.6,
-            current_score: 0.4,
-            delta: -0.20,
-            severity: RegressionSeverity::Severe,
-        },
-    ];
-    let d = decide(&cfg, &baseline, &post, &regressions);
-    match d {
-        ImprovementDecision::Revert { reason } => {
-            assert!(reason.contains("specificity"));
-            assert!(reason.contains("temporal_awareness"));
-        }
-        ImprovementDecision::Commit { .. } => {
-            panic!("expected revert with multiple severe regressions")
-        }
-    }
-}
-
-#[test]
-fn find_weak_dimensions_threshold_zero_all_pass() {
-    let score = make_score(0.01);
-    let weak = find_weak_dimensions(&score, 0.0, None);
-    assert!(weak.is_empty(), "threshold 0.0 should pass all dimensions");
-}
-
-#[test]
-fn find_weak_dimensions_threshold_one_all_fail() {
-    let score = make_score(0.99);
-    let weak = find_weak_dimensions(&score, 1.0, None);
-    assert_eq!(weak.len(), 5, "threshold 1.0 should flag all dimensions");
-}
-
-#[test]
-fn summarize_cycle_negative_improvement() {
-    let cycle = ImprovementCycle {
-        baseline: make_score(0.80),
-        proposed_changes: vec![],
-        post_score: Some(make_score(0.70)),
-        regressions: vec![],
-        decision: Some(ImprovementDecision::Revert {
-            reason: "net negative".into(),
-        }),
-        final_phase: ImprovementPhase::Decide,
-        weak_dimensions: Vec::new(),
-        weak_dimension_details: Vec::new(),
-        target_dimension: None,
-    };
-    let summary = summarize_cycle(&cycle);
-    assert!(summary.contains("REVERT"));
-    assert!(summary.contains("-10.0%"));
-}
-
-#[test]
-fn find_weak_dimensions_sorted_by_deficit() {
-    let score = GymSuiteScore {
-        suite_id: "test".to_string(),
-        overall: 0.65,
-        dimensions: ScoreDimensions {
-            factual_accuracy: 0.55,       // deficit 0.05
-            specificity: 0.50,            // deficit 0.10
-            temporal_awareness: 0.70,     // above threshold
-            source_attribution: 0.30,     // deficit 0.30 (largest)
-            confidence_calibration: 0.58, // deficit 0.02
-        },
-        scenario_count: 6,
-        scenarios_passed: 6,
-        pass_rate: 1.0,
-        recorded_at_unix_ms: None,
-    };
-    let weak = find_weak_dimensions(&score, 0.6, None);
-    assert_eq!(weak.len(), 4);
-    // Sorted by deficit descending
-    assert_eq!(weak[0].name, "source_attribution");
-    assert!((weak[0].deficit - 0.30).abs() < 1e-9);
-    assert_eq!(weak[1].name, "specificity");
-    assert_eq!(weak[2].name, "factual_accuracy");
-    assert_eq!(weak[3].name, "confidence_calibration");
-    assert!(weak[0].deficit >= weak[1].deficit);
-    assert!(weak[1].deficit >= weak[2].deficit);
-    assert!(weak[2].deficit >= weak[3].deficit);
-}
-
-#[test]
-fn summarize_cycle_shows_deficit_when_details_present() {
-    use super::types::WeakDimension;
-    let cycle = ImprovementCycle {
-        baseline: make_score(0.50),
-        proposed_changes: vec![],
-        post_score: None,
-        regressions: vec![],
-        decision: None,
-        final_phase: ImprovementPhase::Analyze,
-        weak_dimensions: vec!["source_attribution".into()],
-        weak_dimension_details: vec![WeakDimension {
-            name: "source_attribution".into(),
-            deficit: 0.25,
-        }],
-        target_dimension: None,
-    };
-    let summary = summarize_cycle(&cycle);
-    assert!(summary.contains("source_attribution (25.0% deficit)"));
-}
-
-#[test]
-fn find_weak_dimensions_all_below_preserves_deficit_sort() {
-    // All five dimensions below threshold — verify sorting by deficit descending.
-    let score = GymSuiteScore {
-        suite_id: "test".to_string(),
-        overall: 0.30,
-        dimensions: ScoreDimensions {
-            factual_accuracy: 0.50,       // deficit 0.10
-            specificity: 0.45,            // deficit 0.15
-            temporal_awareness: 0.40,     // deficit 0.20
-            source_attribution: 0.30,     // deficit 0.30
-            confidence_calibration: 0.55, // deficit 0.05
-        },
-        scenario_count: 6,
-        scenarios_passed: 6,
-        pass_rate: 1.0,
-        recorded_at_unix_ms: None,
-    };
-    let weak = find_weak_dimensions(&score, 0.60, None);
-    assert_eq!(weak.len(), 5);
-    // Verify strictly descending deficit order
-    for window in weak.windows(2) {
-        assert!(
-            window[0].deficit >= window[1].deficit,
-            "{} deficit ({}) should be >= {} deficit ({})",
-            window[0].name,
-            window[0].deficit,
-            window[1].name,
-            window[1].deficit,
-        );
-    }
-    assert_eq!(weak[0].name, "source_attribution");
-    assert_eq!(weak[4].name, "confidence_calibration");
-}
-
-#[test]
-fn find_weak_dimensions_at_exact_threshold_not_weak() {
-    // Dimensions exactly at threshold should NOT be flagged as weak.
-    let score = GymSuiteScore {
-        suite_id: "test".to_string(),
-        overall: 0.60,
-        dimensions: ScoreDimensions {
-            factual_accuracy: 0.60,
-            specificity: 0.60,
-            temporal_awareness: 0.60,
-            source_attribution: 0.60,
-            confidence_calibration: 0.60,
-        },
-        scenario_count: 6,
-        scenarios_passed: 6,
-        pass_rate: 1.0,
-        recorded_at_unix_ms: None,
-    };
-    let weak = find_weak_dimensions(&score, 0.60, None);
-    assert!(
-        weak.is_empty(),
-        "dimensions at exact threshold should not be weak"
-    );
-}
-
-#[test]
-fn summarize_cycle_multiple_weak_dimension_details() {
-    use super::types::WeakDimension;
-    let cycle = ImprovementCycle {
-        baseline: make_score(0.40),
-        proposed_changes: vec![],
-        post_score: None,
-        regressions: vec![],
-        decision: None,
-        final_phase: ImprovementPhase::Analyze,
-        weak_dimensions: vec!["source_attribution".into(), "specificity".into()],
-        weak_dimension_details: vec![
-            WeakDimension {
-                name: "source_attribution".into(),
-                deficit: 0.30,
-            },
-            WeakDimension {
-                name: "specificity".into(),
-                deficit: 0.15,
-            },
-        ],
-        target_dimension: None,
-    };
-    let summary = summarize_cycle(&cycle);
-    assert!(summary.contains("source_attribution (30.0% deficit)"));
-    assert!(summary.contains("specificity (15.0% deficit)"));
-}
-
-#[test]
-fn summarize_cycle_lists_proposed_change_count() {
-    // Happy path: when proposed_changes is non-empty, summary includes a count line.
-    let cycle = ImprovementCycle {
-        baseline: make_score(0.70),
-        proposed_changes: vec![
-            ProposedChange {
-                file_path: "src/a.rs".into(),
-                description: "tighten error handling".into(),
-                expected_impact: "fewer regressions".into(),
-            },
-            ProposedChange {
-                file_path: "src/b.rs".into(),
-                description: "expand docs".into(),
-                expected_impact: "clearer behaviour".into(),
-            },
-            ProposedChange {
-                file_path: "src/c.rs".into(),
-                description: "split helper".into(),
-                expected_impact: "easier testing".into(),
-            },
-        ],
-        post_score: Some(make_score(0.75)),
-        regressions: vec![],
-        decision: Some(ImprovementDecision::Commit {
-            net_improvement: 0.05,
-        }),
-        final_phase: ImprovementPhase::Decide,
-        weak_dimensions: Vec::new(),
-        weak_dimension_details: Vec::new(),
-        target_dimension: None,
-    };
-    let summary = summarize_cycle(&cycle);
-    assert!(
-        summary.contains("Proposed changes: 3"),
-        "summary should include count of proposed changes, got: {summary}"
-    );
-}
-
-#[test]
-fn summarize_cycle_lists_weak_dimensions_without_details() {
-    // Edge case: when weak_dimension_details is empty but weak_dimensions is
-    // populated (e.g. legacy cycles deserialized from JSON predating the
-    // weak_dimension_details field), summary falls back to the names-only line.
-    let cycle = ImprovementCycle {
-        baseline: make_score(0.50),
-        proposed_changes: vec![],
-        post_score: None,
-        regressions: vec![],
-        decision: None,
-        final_phase: ImprovementPhase::Analyze,
-        weak_dimensions: vec!["specificity".into(), "source_attribution".into()],
-        weak_dimension_details: Vec::new(),
-        target_dimension: None,
-    };
-    let summary = summarize_cycle(&cycle);
-    assert!(
-        summary.contains("Weak dimensions: specificity, source_attribution"),
-        "summary should fall back to the names-only weak dimensions line, got: {summary}"
-    );
-    // Detail format must NOT appear because details are empty.
-    assert!(!summary.contains("% deficit"));
-}
-
-#[test]
-fn summarize_cycle_omits_weak_dimensions_when_both_lists_empty() {
-    // Edge case: when both weak_dimension_details AND weak_dimensions are empty,
-    // summary must not include any "Weak dimensions:" line at all.
-    let cycle = ImprovementCycle {
-        baseline: make_score(0.90),
-        proposed_changes: vec![],
-        post_score: Some(make_score(0.92)),
-        regressions: vec![],
-        decision: Some(ImprovementDecision::Commit {
-            net_improvement: 0.02,
-        }),
-        final_phase: ImprovementPhase::Decide,
-        weak_dimensions: Vec::new(),
-        weak_dimension_details: Vec::new(),
-        target_dimension: None,
-    };
-    let summary = summarize_cycle(&cycle);
-    assert!(
-        !summary.contains("Weak dimensions"),
-        "summary must not include weak dimensions line when none exist, got: {summary}"
-    );
-}
-
-// Tests previously inlined in src/self_improve/cycle.rs (#1266 burndown)
-mod cycle_inline {
-    use super::super::cycle::*;
-    use super::super::types::{
-        ImprovementConfig, ImprovementCycle, ImprovementDecision, ImprovementPhase,
-    };
-    use crate::gym_bridge::ScoreDimensions;
-    use crate::gym_scoring::{GymSuiteScore, Regression, RegressionSeverity};
-
-    fn make_score(overall: f64, dims: ScoreDimensions) -> GymSuiteScore {
-        GymSuiteScore {
-            suite_id: "test-suite".to_string(),
-            overall,
-            dimensions: dims,
-            scenario_count: 5,
-            scenarios_passed: 5,
-            pass_rate: 1.0,
-            recorded_at_unix_ms: None,
-        }
-    }
-
-    fn dims(fa: f64, sp: f64, ta: f64, sa: f64, cc: f64) -> ScoreDimensions {
-        ScoreDimensions {
-            factual_accuracy: fa,
-            specificity: sp,
-            temporal_awareness: ta,
-            source_attribution: sa,
-            confidence_calibration: cc,
-        }
-    }
-
-    #[test]
-    fn decide_commit_when_net_positive() {
-        let config = ImprovementConfig {
-            min_net_improvement: 0.01,
-            max_single_regression: 0.1,
-            ..ImprovementConfig::default()
-        };
-        let baseline = make_score(0.70, dims(0.7, 0.7, 0.7, 0.7, 0.7));
-        let post = make_score(0.75, dims(0.75, 0.75, 0.75, 0.75, 0.75));
-        let decision = decide(&config, &baseline, &post, &[]);
-        assert!(matches!(decision, ImprovementDecision::Commit { .. }));
-    }
-
-    #[test]
-    fn decide_revert_below_threshold() {
-        let config = ImprovementConfig {
-            min_net_improvement: 0.10,
-            max_single_regression: 0.5,
-            ..ImprovementConfig::default()
-        };
-        let baseline = make_score(0.70, dims(0.7, 0.7, 0.7, 0.7, 0.7));
-        let post = make_score(0.72, dims(0.72, 0.72, 0.72, 0.72, 0.72));
-        let decision = decide(&config, &baseline, &post, &[]);
-        assert!(matches!(decision, ImprovementDecision::Revert { .. }));
-    }
-
-    #[test]
-    fn decide_revert_on_severe_regression() {
-        let config = ImprovementConfig {
-            min_net_improvement: 0.01,
-            max_single_regression: 0.05,
-            ..ImprovementConfig::default()
-        };
-        let baseline = make_score(0.70, dims(0.7, 0.7, 0.7, 0.7, 0.7));
-        let post = make_score(0.75, dims(0.8, 0.8, 0.8, 0.8, 0.8));
-        let regression = Regression {
-            dimension: "specificity".to_string(),
-            baseline_score: 0.9,
-            current_score: 0.7,
-            delta: -0.2,
-            severity: RegressionSeverity::Severe,
-        };
-        let decision = decide(&config, &baseline, &post, &[regression]);
-        assert!(matches!(decision, ImprovementDecision::Revert { .. }));
-    }
-
-    #[test]
-    fn find_weak_dimensions_below_threshold() {
-        let score = make_score(0.6, dims(0.3, 0.8, 0.4, 0.9, 0.2));
-        let weak = find_weak_dimensions(&score, 0.5, None);
-        assert_eq!(weak.len(), 3);
-        let names: Vec<&str> = weak.iter().map(|w| w.name.as_str()).collect();
-        assert!(names.contains(&"factual_accuracy"));
-        assert!(names.contains(&"temporal_awareness"));
-        assert!(names.contains(&"confidence_calibration"));
-    }
-
-    #[test]
-    fn find_weak_dimensions_with_target_filter() {
-        let score = make_score(0.6, dims(0.3, 0.8, 0.4, 0.9, 0.2));
-        let weak = find_weak_dimensions(&score, 0.5, Some("specificity"));
-        assert!(weak.is_empty()); // specificity is 0.8, above threshold
-    }
-
-    #[test]
-    fn summarize_cycle_baseline_only() {
-        let cycle = ImprovementCycle {
-            baseline: make_score(0.75, ScoreDimensions::default()),
-            proposed_changes: Vec::new(),
-            post_score: None,
-            regressions: Vec::new(),
-            decision: Some(ImprovementDecision::Revert {
-                reason: "no changes".to_string(),
-            }),
-            final_phase: ImprovementPhase::Analyze,
-            weak_dimensions: Vec::new(),
-            weak_dimension_details: Vec::new(),
-            target_dimension: None,
-        };
-        let summary = summarize_cycle(&cycle);
-        assert!(summary.contains("Baseline:"));
-        assert!(summary.contains("REVERT"));
-        assert!(!summary.contains("Post-change:"));
-    }
-
-    #[test]
-    fn summarize_cycle_with_post_score() {
-        let cycle = ImprovementCycle {
-            baseline: make_score(0.70, ScoreDimensions::default()),
-            proposed_changes: vec![],
-            post_score: Some(make_score(0.80, ScoreDimensions::default())),
-            regressions: Vec::new(),
-            decision: Some(ImprovementDecision::Commit {
-                net_improvement: 0.10,
-            }),
-            final_phase: ImprovementPhase::Decide,
-            weak_dimensions: Vec::new(),
-            weak_dimension_details: Vec::new(),
-            target_dimension: Some("specificity".to_string()),
-        };
-        let summary = summarize_cycle(&cycle);
-        assert!(summary.contains("Post-change:"));
-        assert!(summary.contains("COMMIT"));
-        assert!(summary.contains("Target dimension: specificity"));
-    }
-}
